requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
groq
huggingface-hub
requests
